Delete src folder. Create index.js and App.js
Description: Delete the original src folder along with everything in it and created my own src folder and add index.js and App.js file in it. Started App.js by typing rafce and import React and ReactDOM in the index.js file. connect the index.js file to the root in index.html file which is in the public folder by typing ReactDOM.render(<App />, document.getElementById('root'))