Imported containers: Typography and IconButton
Typography container will have a variant of h2 and a color of textsecondary. Specifying the text and background colors of secondary content means that the browser will not override those colors. With this technique the author uses the default text color and background color of the main area. As a result the colors are completely determined by the user agent via the user's color preferences. The dot notation access the property of product which is description, right now this is an example of using props in reactjs. The <CardActions> container is going to have a prop called disableSpacing witha className of classes.CardActions. Inside the <CardActions> container there is an <IconButton> container which has an attribute called aria-label. Aria-label allows us to specify a string to be used as the accessible label. This overrides any other native labeling mechanism, such as a label element — for example, if a button has both text content and an aria-label , only the aria-label value will be used. The aria-label will display 'Add to Cart' which will override any other native labeling mechanism.
<AddShoppingCart /> container is added which was imported from @material-ui/core. The AddShoppingCart is an icon.